Added pathfinding for the units.
This includes dynamic reshaping of the navigation polygon to avoid the object instances, however there is a bug where if the units slide into each other, they slow down to the slowest unit amongst the colliding group, even though they are designed to speed up when colliding. This will need to be fixed.